tests/wpt/css-tests/css-backgrounds-3_dev/xhtml1print/background-origin-001.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <title>CSS Backgrounds and Borders Test: background-origin - initial and supported values</title>
    <link href="http://www.intel.com" rel="author" title="Intel" />
    <link href="http://www.w3.org/TR/css3-background/#the-background-origin" rel="help" />
    <meta content="dom" name="flags" />
    <meta content="Check if background-origin initial value is padding-box and supports values border-box, padding-box and content-box" name="assert" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <div id="test"></div>
    <script>
        test(function() {
            assert_equals(getComputedStyle(document.getElementById("test"), null).getPropertyValue("background-origin"),
                "padding-box", "background-origin initial value");
        }, "background-origin_initial");

        document.getElementById("test").style.backgroundOrigin = "border-box";
        test(function() {
            assert_equals(getComputedStyle(document.getElementById("test"), null).getPropertyValue("background-origin"),
                "border-box", "background-origin supporting value");
        }, "background-origin_border-box");

        document.getElementById("test").style.backgroundOrigin = "padding-box";
        test(function() {
            assert_equals(getComputedStyle(document.getElementById("test"), null).getPropertyValue("background-origin"),
                "padding-box", "background-origin supporting value");
        }, "background-origin_padding-box");

        document.getElementById("test").style.backgroundOrigin = "content-box";
        test(function() {
            assert_equals(getComputedStyle(document.getElementById("test"), null).getPropertyValue("background-origin"),
                "content-box", "background-origin supporting value");
        }, "background-origin_content-box");
    </script>
  

</body></html>
